make auto id test stable
